kes  Make bscan ignore ACL streams. kes  Update projects files with new Feature Requests. kes  Generate current English and French manuals and upload      them to the site. kes  Update the home page to include the new British/French flags to      switch between translations. Work done by Alexandre Baron. kes  Make disk-changer script print an error message if the changer      directory is not defined. kes  Create a generic DB routine for getting an int (32/64 bit) from      the DB. kes  Add more migration code. kes  Add a few more files to .cvsignore here and there.
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@3201 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/dird/dird_conf.h
@@ -463,8 +463,8 @@ public:
    uint32_t MaxVolFiles;              /* Maximum files on the Volume */
    uint64_t MaxVolBytes;              /* Maximum bytes on the Volume */
    utime_t MigrationTime;             /* Time to migrate to next pool */
-   uint32_t MigrationHighBytes;       /* When migration starts */
-   uint32_t MigrationLowBytes;        /* When migration stops */
+   uint64_t MigrationHighBytes;       /* When migration starts */
+   uint64_t MigrationLowBytes;        /* When migration stops */
    POOL  *NextPool;                   /* Next pool for migration */
    alist *storage;                    /* Where is device -- list of Storage to be used */
    bool  use_catalog;                 /* maintain catalog for media */
